+++ CHANGELOG.md
@@ -42,8 +42,12 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Fixed
 
+- Reinstallation using the installation script on systems where AdGuard Home
+  is the only DNS server ([#4809]).
+
 - Blocked requests without an EDNS(0) OPT record ([#8183]).
 
+[#4809]: https://github.com/AdguardTeam/AdGuardHome/issues/4809
 [#8183]: https://github.com/AdguardTeam/AdGuardHome/issues/8183
 
 <!--